+++ default.nix
@@ -3,7 +3,7 @@
 }:
 mkDerivation {
   pname = "polysemy-mocks";
-  version = "0.2.0.0";
+  version = "0.3.0.0";
   src = ./.;
   libraryHaskellDepends = [ base polysemy template-haskell ];
   testHaskellDepends = [ base hspec polysemy ];